requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
flower==2.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.6
kombu==5.3.7
numpy==1.24.4
openai==1.30.5
opencv-python==4.10.0.82
pandas==2.0.3
pgvector==0.2.5
pillow==10.3.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-core==2.16.3
PyJWT==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tornado==6.4
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
